requirements.txt
Django>=2.0,<2.1
#psycopg2
django-compressor
django-oscar>=2.0,<2.1
django-oscar-promotions
django-tables2
django-modelcluster
wagtail>=2.1,<2.2
django-taggit
django-libsass
libsass
Pillow
pycountry
sorl-thumbnail
#Django>=1.8,<1.9
#psycopg2==2.6
#django-compressor>=1.4
#django-oscar>=1.2,<1.3
#django-modelcluster>=0.6
#wagtail>=1.5,<1.6
#django-taggit>=0.14.0
#django-libsass>=0.2
#libsass==0.8.2
#Pillow==2.8.1
#pycountry==1.20

# Dependencies for hosting
django-redis==3.8.2
elasticsearch==1.2.0

# Dependencies for Heroku deployment
#dj-database-url==0.3.0
#whitenoise==2.0.4
#uwsgi==2.0.11.2
